+++ js/main.js
@@ -1,4 +1,2 @@
 import { createData } from './create-data.js';
 const data = createData();
-
-console.log(data);
